Added dependency on CocoaLumberjack for all our logging.
Changed our NSLog override and DLog and ELog to use CocoaLumberjack.
These macros are slightly different to the DDLog ones -- they assume
a fixed log level of INFO in release builds and DEBUG in debug builds,
so that there is no need to define ddLogLevel at the top of the file.

They also set NSLog and DLog to the C version of the macro, rather
than the ObjC version, so that they can be used in static contexts.
This is for compat with existing usage.

There is now NSLogError, NSLogWarn, and NSLogInfo which are using the
ObjC versions.