Add InlineTimingHeap*.
These are the same as the existing InlineTiming macros, except they
store the state in a heap-allocated object (an instance of InlineTiming)
rather than stack-allocated arrays.

This will be slightly slower, and it means that you need to pass a
reference to the object each time, but it means that they can be used
between blocks or function calls.