Change the PerUser interface to have getTipbitUser, setTipbitUserAndSynchronize:, and clearTipbitUserAndSynchronize:.
This means that kTipbitUser is private to this module now.  It
also means that we're not dependent on callers to remember to
match a removeObjectForKey:kTipbitUser call with the correct
clearTipbitUserCache call -- we can manage the cache internally.

It also means that a setTipbitUserAndSynchronize followed by getTipbitUser
will use the cache and won't go back to NSUserDefaults first time.