Rewrite to avoid a retain cycle.
The call to NSInvocationOperation.initWithTarget makes an
NSInvocation instance that retains its arguments, and assigns it to
self.invocation.  There is no way to release these arguments again,
since self.invocation is readonly and setting self.invocation.target = nil
doesn't release target (even though it clearly should do).  In our
case, the target was self, which caused a retain cycle.

Get around this by having a separate helper class.  This breaks the cycle
because now the helper class is the target of the invocation.

Discard the block once it's been invoked, which should
reduce the chance of further retain cycles if our caller isn't very careful.