Performance improvements for NSDate+ISO8601 parsing.
Add CBLParseDate from CouchbaseLite,
https://github.com/couchbase/couchbase-lite-ios, d05943f1b03.
This is Copyright (c) 2012-2013 Couchbase, Inc, and licensed using the
Apache Liense 2.0.  That code in turn includes code from SQLite, which is
in the public domain, and credits Astronomical Algorithms by Jean Meeus.
The CBL code also credits @nirvana and @AnuragMishra for earlier work.

Use CBLParseISO8601Date to replace our use of NSDateFormatter in
[NSDate dateFromIso8601].  This implementation is 21.7x faster than
using a single NSDateFormatter in a loop, and 756x faster than our previous
implementation (because we were creating the NSDateFormatter each time for
thread-safety reasons).  Yes, that's 756 *times* faster.

Also, add [NSDate timeIntervalSinceReferenceDateFromIso8601].  This is
even faster than [NSDate dateFromIso8601], because it doesn't need to allocate
an NSDate.