Added a script to run gcovr.
This needs gcovr-3.0.tar.gz to be placed in the same directory as
this script (something we ensure in our Jenkins config).

Call it with something like

./run-gcovr.sh -e tidbits.TidbitsTest

You will get Cobertura-style coverage info in coverage.xml,
which you can get Jenkins to render.

You might also consider --gcov-filter='.*#mypackage#.*'

See http://gcovr.com for gcovr.

gcov support is completely broken in Xcode 5 -- see
google-toolbox-for-mac/UnitTesting/ for the fix for that.