Replace NSInputStream.writeToFile with writeToFileAndNSData.
This will return the NSData that was put in the file.  If the
file can't be written (e.g. because the screen is locked) then
you'll get the data but without the file being created.

As part of this, change NSData.dataWithContentsOfStream to take
an NSError out parameter and improve its error handling throughout.