Add TBCAssertAllBuilds.
This is in all builds (not just DEBUG builds).  It acts similarly
to NSCAssert, except it also uses Breadcrumbs to save a marker in
the crash dump, and uses NSLogError and flushes the DDLog so that
we have a good chance that it gets into the log file too, before
the app is terminated.