Add TBJUnitTestObserver, a new subclass of XCTestObserver that writes a JUnit report at the end of the test.
Until now we've been using a Jenkins plugin to parse the console
output to get the test info, but that has a problem where it
gets out of sync if Lumberjack's logging gets interleaved with XCTestLog's.
This avoids parsing the console at all.

This also captures the Lumberjack logging for the duration of the
test and puts that in the report too.  The Jenkins plugin doesn't do this.