Fix TBUserDefaults when loading a plist with NSFileProtection set, but the screen is locked.
In this case, we would default the settings to the empty dictionary,
because we couldn't read the plist, but that's wrong because it
causes all those settings to be reset to their defaults.  They
would then be lost once the screen unlocked, because the next
time the settings were written we'd overwrite with the blank dictionary.