Add SYNTHESIZE_ASSOCIATED_OBJ.
This is a replacement for SYNTHESIZE_ASSOCIATED_OBJECT where we use the
selector for the getter as the key, thus avoiding the need for the caller
to pass a key in.

This is better than SYNTHESIZE_ASSOCIATED_OBJECT in all situations.  The only
reason that's still alive is for backwards compat.